Upgrade nrf51-hal to 0.12.1
This is a large change down to the nrf51-hal API changing significantly
between 0.7 and 0.10. Below I cover the big changes from the perspective
of what has changed in the nrf51-hal API first as that is where I
started.

The library
===========

`nrf51_hal::timer`
------------------
This is the biggest change. The new timer interface is significantly
stripped back so all of the interrupt and compare config in
`microbit::display::timer` now have to be done directly on the PAC.
A gotcha I hit here was that on the nrf51 all but the first timer seem
to only be 16 bit. While the implementation has changed significantly
here the public interface has not changed.

`nrf51_hal::gpio`
-----------------
How the pins are accessed and how they are named has changed however at
the end of the day, they're still GPIO pins so it's not too difficult.

`nrf51_hal::twi`
----------------
The name has changed, arguments are slightly different and you have to
pass the frequency explicitly now. Other than that it seems the same.

`nrf51_hal::uart`
-----------------
The name has changed, arguments are slightly different and you have to
pass whether the parity bit is included explicitly now. There doesn't
seem to be a way to split it into separate read and write parts but for
the examples that didn't seem to be a problem.

The examples
============

I have updated all examples to compile and tried to run them however not
all of them work.

Fully working
-------------
- `gpio_direct_blinky.rs`
- `gpio_hal_blinky.rs`
- `gpio_hal_ledbutton.rs`
- `gpio_hal_printbuttons.rs`
- `led_blocking.rs`
- `led_nonblocking.rs` (it's a bit flickery but it was before, I'll look
  at that separately
- `led_rtfm.rs`
- `rng_direct_printrngserial.rs`
- `serial_direct_echo.rs`
- `serial_direct_helloworld.rs`
- `serial_hal_blocking_echo.rs`

Not working
-----------

None of these work for me on this branch or on master. I would
appreciate someone who has them working on master testing them for me.

- `rng_hal_printrandserial.rs`: crashes on line 84 at `rng.next_u32()`.
- `gpio_hal_receivedcf77.rs`: just get a zero timestamp.

I believe all these are failing because I have v1.5 see https://github.com/therealprof/microbit/issues/9
- `i2c_direct_printmagserial.rs`
- `i2c_hal_printmagserial.rs`
- `i2c_haldriver_printmagserial.rs`