Move led pins into separate struct
The main goal here is to make it easier to support the v2 board which
has completely different pin layout as the interface will still be
similar.

As an added benefit it makes the constructor for `led::Display` a bit
clearer as rather than a long list of positional arguments you
effectively provide named arguments. This is a pattern I've seen used
in the nrf-hal.