Refactor display to take DisplayPins
This change makes use of the approach from @mattheww in
mattheww/rmicrobit where the display method (`initialise_display` and
`handle_display_event`) take a `DisplayPins` struct rather than a
reference to the GPIO peripheral. This struct is never used but ensures
nothing else is accidentally using the pins it uses. It then modifies
the pins directly on the peripheral so that multiple can be updated
together.

I have moved and renamed the `led::Pins` struct as it seems to make more
sense in the `gpio` module as it's used in `led` and `display` now.

Finally, I have introduced a `display_pins` macro, mostly because I was
getting tired of copying around the initialization of the big pin
struct.