Add support for micro:bit v2
Add features for microbit-v1 and microbit-v2

Add microbit v2 support to display module
-----------------------------------------

This was hard and I am not super happy with the solution. It is much
more bit twiddling than I am comfortable with.

The issue that made it so difficult is that on the microbit v2 the
column pins are on the P0 and P1 ports so all the row wise writes have
to be split accross the two ports.

Get microbit v2 led examples working
------------------------------------

- Create a new v2 gpio module and conditionally use v1 or v2
- Move NUM_ROWS and NUM_COLS consts into gpio module so they can differ
  between v1 and v2
- Move degrading the LED pins into the gpio module so it can differ
  between v1 and v2

The code for the nonblocking display is a nightmare. The columns are
split between the two GPIO ports so the cols argument to
`display_row_leds` needs to be split between the two ports. This is a
bit painful but doable. The bit that is really confusing me is that
columns 2 and 3 seem to be swapped.