Update nrf-hal dependencies
The unreleased fix is now released in 0.14.0 so pulling 0.13.0 from git
is no longer needed (and no longer works without pinning to a commit).

However, this changes the cortex-m version from 0.6.x to 0.7.3 which
causes problems for some of the examples where the old cortex-m is still
brought in by other dependencies. The fix for the display-rtic example
is to use cortex-m-7 feature until cortex-m-rtic 0.6 is released. For
rubble the only fix I could find was to grab the git version as it does
use the latest nrf5x pac there it's just not released yet.